% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcolaEnhancedFootPrintTable.pat.process}
\alias{calcolaEnhancedFootPrintTable.pat.process}
\title{calcola EFT}
\usage{
calcolaEnhancedFootPrintTable.pat.process(
  dataLoaderOBJ,
  skip.less.than = 0,
  threshold.perc = 1e-08
)
}
\description{
Funzione per calcolare la EFM
}
