% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcolaFootPrintTable.pat.process}
\alias{calcolaFootPrintTable.pat.process}
\title{calcola FT}
\usage{
calcolaFootPrintTable.pat.process(
  dataLoaderOBJ,
  skip.less.than = 0,
  threshold.perc = 1e-08
)
}
\description{
Funzione per calcolare la FM
}
