% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careFlowMiner.R
\name{careFlowMiner}
\alias{careFlowMiner}
\title{A class for a revisited Careflow Mining}
\usage{
careFlowMiner(verbose.mode = FALSE)
}
\description{
This is an implementation of the Care Flow Mining algorithm, a bit revisited
}
