% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careFlowMiner.wrapper.R
\name{CFM.loadDataset}
\alias{CFM.loadDataset}
\title{CFM: load a dataset}
\usage{
CFM.loadDataset(objCFM, inputData, dateToColumnName = NA, dateToFormat = "")
}
\arguments{
\item{objCFM}{the \code{careFlowMiner} obj.}

\item{inputData}{the data to load formattes as a dataLoader$getData() structure}

\item{dateToColumnName}{the name of the column containing the optional \code{dateTo} of the clinical event}

\item{dateToFormat}{the format of the optional \code{dateTo} ( this because \code{dateFrom} and \code{dateTo} can have different formats)}
}
\description{
Load a dataLoader$getData() structure
}
\examples{
\dontrun{

# instantiate a dataLoader object
library(pMineR)

tmp.objDL <- dataLoader()
invisible(tmp.objDL$load.data.frame( mydata = data.frame(myInputCSVData) , IDName = "PatientID",EVENTName = "clinicalEvent",
 dateColumnName = "STRT_DT", format.column.date = "\%Y-\%m-\%d \%H:\%M:\%S" ))
 DLS <- tmp.objDL$getData()

a <- careFlowMiner()
a$loadDataset(DLS = DLS , dateToColumnName = "END_DT",dateToFormat="\%Y-\%m-\%d \%H:\%M:\%S") 

}
}
