% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careFlowMiner.wrapper.R
\name{CFM.pathBeetweenStackedNodes}
\alias{CFM.pathBeetweenStackedNodes}
\title{CFM: pack start and end nodes and plot the paths}
\usage{
CFM.pathBeetweenStackedNodes(
  objCFM,
  fromState,
  toState,
  stratifyFor = "",
  minPath = FALSE,
  stratificationValues = c(),
  fisher.threshold = 0.1,
  kindOfGraph = "dot",
  arcColor = "black",
  arc.fontsize = 10,
  arc.fontcolor = "red",
  arr.States.color = c(),
  set.to.gray.color = "WhiteSmoke",
  p.value.threshold = 0.05
)
}
\arguments{
\item{objCFM}{the \code{careFlowMiner} obj.}
}
\description{
stack two given nodes, respectively start and end, and show all the paths between them. It also allow to stratify for a given covariate (optional)
}
