% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careFlowMiner.wrapper.R
\name{CFM.plotCFGraphComparison}
\alias{CFM.plotCFGraphComparison}
\title{CFM: Compare two Graphs}
\usage{
CFM.plotCFGraphComparison(
  objCFM,
  stratifyFor,
  stratificationValues,
  depth = 4,
  fisher.threshold = 0.1,
  arr.States.color = c()
)
}
\arguments{
\item{objCFM}{the \code{careFlowMiner} obj.}

\item{stratifyFor}{the dicotomic attribute you want to use to stratify. It should be the same for all the record of a specific patients (it is a baseline attribute).}

\item{depth}{max depth level (default = 4)}

\item{fisher.threshold}{bla bla bla}

\item{arr.States.color}{an array contaning the association between nodes and color}
}
\description{
Compare two graphs, stratyfing for a given covariate
}
