% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confCheck_easy.R
\name{confCheck_easy}
\alias{confCheck_easy}
\title{A simple conformance checking class}
\usage{
confCheck_easy(verbose.mode = TRUE)
}
\arguments{
\item{verbose.mode}{boolean. If TRUE some messages will appear in console, during the computation; otherwise the computation will be silent.}
}
\description{
A first module for making conformance checking
}
