% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoader.R
\name{dataLoader}
\alias{dataLoader}
\title{Load the event-logs}
\usage{
dataLoader(
  verbose.mode = TRUE,
  max.char.length.label = 50,
  save.memory = FALSE
)
}
\arguments{
\item{verbose.mode}{boolean. If TRUE some messages will appear in console, during the computation; otherwise the computation will be silent.}

\item{max.char.length.label}{numeric. It defines the max length of the event name strings}

\item{save.memory}{boolean. If TRUE, dataLoader() avoid to keep in memory the entire original csv.}
}
\description{
A loader for csv based log files. It also calculates the footprint table, transition matrix probabilities, and presents data in different shapes. The public methods are:
             \itemize{
             \item \code{dataLoader() } the costructor
             \item \code{load.csv( ... ) } loads the csv file into the \code{dataLoader} object
             \item \code{load.data.frame() } loads a data.frame into the \code{dataLoader} object
             \item \code{getData() } return the processed, previously-loaded, data
             \item \code{addDictionary() } add a dictionary in order, afterward, to translate or group some event name
             } 
             You can use the cited methods directly on the object, (they are made available as element of the object list) or can be invocked using the 'wrapping function', with the prefix 'DL.'
             
             The consturctor admit the following parameters:
verbose.mode are some notification wished, during the computation? The defaul value is \code{true}
}
