% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand.grid.unique}
\alias{expand.grid.unique}
\title{Some useful tools}
\usage{
expand.grid.unique(x, y, include.equals = FALSE)
}
\description{
A class which provide some tools. pMineR intarnal use only.
}
