% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotCohortTimeline}
\alias{plotCohortTimeline}
\title{A function to plot nice timeline}
\usage{
plotCohortTimeline(
  objDL.obj,
  lst.to.join = list(),
  lst.spike = list(),
  UM = "days",
  y.grid = NA,
  ordered = FALSE,
  decreasing = TRUE,
  y.grid.col = "lightgrey",
  x.grid.col = "lightgrey"
)
}
\description{
wow
}
